requirements.txt
torch
transformers>=4.44.0
accelerate>=0.33.0
safetensors>=0.4.3
diffusers
huggingface_hub>=0.25.0
pillow>=10.2.0
numpy>=1.26.0
